pyproject.toml
[project]
name = "dq-rag-specs"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
  "pandas>=1.5,<2.1", "pyarrow>=10.0,<15.0", "pandera>=0.14,<0.18",
  "pydantic>=1.10,<2.0", "fastapi>=0.100,<0.110", "uvicorn[standard]>=0.20,<0.25",
  "pypdf>=3.0,<4.0", "faiss-cpu>=1.7,<1.8", "sentence-transformers>=2.2,<3.0",
  "python-dotenv>=0.19"
]
[project.optional-dependencies]
dev = ["pytest>=6.0,<8.0", "ruff>=0.0.280,<0.6", "black>=22.0,<25.0", "mypy>=0.991,<1.8", "pre-commit>=2.0,<4.0"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
)/
'''